# docs/api.yaml
openapi: 3.0.0
info:
  title: Treehole Chat API
  version: 1.0.0
  description: |
    面向树洞聊天业务的作业接口（认证 + 帖子 CRUD + 分页筛选）。
    使用前请先配置 backend/.env 中 DB_HOST/DB_PORT/DB_USER/DB_PASSWORD/DB_NAME，
    并确保数据库已执行 backend/sql/init.sql。
servers:
  - url: http://localhost:3000
paths:
  /api/auth/register:
    post:
      summary: 注册用户（映射到 users）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
                displayName: { type: string }
      responses:
        '201':
          description: 注册成功
        '400':
          description: 参数错误
  /api/auth/login:
    post:
      summary: 登录（使用注册后的用户名和密码）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: 登录成功
        '401':
          description: 认证失败
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer